Formally separate React components into “components” and “containers”. Document in README.
